Update pkg/webhook/vault_secret_manager.go
There is a [workaround](https://www.vaultproject.io/docs/platform/k8s/injector/examples#environment-variable-example) which involves mounting a template formatted file that contains `export API_KEY="{{ .Data.data.api_key }}"` and then sourcing this file as an extra step. But unless the user takes this extra sourcing step, this is still file mounting. So I would go with this Error message since the user should be warned that the expected result from requesting Env var will not be achieved with this.

Co-authored-by: Haytham Abuelfutuh <haytham@afutuh.com>
Signed-off-by: Tim Bauer <tim.bauer@wolt.com>